Split web/direct.go into separate files
